==> code/step_1_import_clean.rmd <==
---
title: "Import and Clean Data"
author: "Jonathan Salas"
date: "Last Updated `r format(Sys.time(), '%d %B, %Y')`"
output: html_document
---

# Start with a clean slate

## Remove all objects

```{r echo = F, include=FALSE}
rm(list = ls(all.names = TRUE)) # Remove all starting code
```

## Detach Packages

```{r include=FALSE}
try(invisible(lapply(paste0('package:', names(sessionInfo()$otherPkgs)), detach, character.only=TRUE, unload=TRUE)), silent = T) # Detach Packages
```

```{r setup, include=FALSE}
library(readxl)
library(janitor)
library(dplyr)
knitr::opts_knit$set(root.dir = "~/R_Projects/VCAS1_prelim_efficiency")
```

# Start with a clean slate

```{r echo = F}
rm(list = ls(all.names = TRUE)) # Remove all starting code
```

# Import the raw data - Point to the correct raw data directory

DONT FORGET TO UPDATE THE INDEX PROCEDURE SHEET!

```{r}
raw_data_dir <- "data/raw/VCAS-rove_time.xlsx"
raw_data_dir_2 <- "data/raw/FormBased_FieldMedicalPFAStudy.xlsx"
```

# Excel import

## Import GLG index procedure data

## Import Data Headers

```{r}
# Return the sheetnames from the excel raw data doc
headers_sheetnames <-
  readxl::excel_sheets(raw_data_dir_2)

# Return the sheetnames from the excel raw data doc
headers_list <-
  lapply(headers_sheetnames, readxl::read_excel, path = raw_data_dir_2, n_max = 2)

# name the dataframes
names(headers_list) <- paste0("headers_",headers_sheetnames)
```

### Remove Empty Rows

```{r}
headers_list <- lapply(headers_list, janitor::remove_empty, "rows")
```

## Import Data Content

```{r}
# Return the sheetnames from the excel raw data doc
dat_sheetnames <-
  readxl::excel_sheets(raw_data_dir_2)

# Return the sheetnames from the excel raw data doc
dat_list <-
  lapply(
    dat_sheetnames,
    readxl::read_excel,
    path = raw_data_dir_2,
    skip = 2,
    na = c(".h", ".b", ".ns", ".n", ".e")
  ) 

# name the dataframes
names(dat_list) <- paste0("dat_",dat_sheetnames)
```

### Remove Empty Rows

```{r}
dat_list <- lapply(dat_list, janitor::remove_empty, "rows")
```

## Import our handwriten rove-time data

### Import Subject Data
```{r}
raw_res <- read_excel(path = raw_data_dir,
                      sheet = "results")
```

### Remove Empty Rows

```{r}
cleaned_res <- janitor::remove_empty(raw_res, which = "rows") |> select(-c(link_to_data, notes))
```

# Save cleaned data

```{r}
save(cleaned_res, 
     dat_list,
     headers_list,
     file = paste0("data/processed/imported_efficiency_stats.RData"))
```

# Version and Package Details

```{r echo=F}
paste(version$version.string, version$nickname)
paste0(
  "RStudio Version ",
  rstudioapi::versionInfo()$version,
  " ",
  rstudioapi::versionInfo()$release_name
) # RStudio Version
subset(data.frame(sessioninfo::package_info()),
       attached == TRUE,
       c(package, loadedversion)) # Loaded Packages
```

# When were these files last rewritten?

```{r}
date()
```
