==> code/step_2_working_data.rmd <==
---
title: "Modify Raw Imported Data: Procedural Analysis Set"
author: "Jonathan Salas"
date: "Last Updated `r format(Sys.time(), '%d %B, %Y')`"
output: html_document
---

# Start with a clean slate

## Remove all objects

```{r echo = F, include=FALSE}
rm(list = ls(all.names = TRUE)) # Remove all starting code
```

## Detach Packages

```{r include=FALSE}
try(invisible(lapply(paste0('package:', names(sessionInfo()$otherPkgs)), detach, character.only=TRUE, unload=TRUE)), silent = T) # Detach Packages
```

```{r setup, include=FALSE}
library(dplyr)
library(janitor)
library(hms)
library(lubridate)
library(stringr)
knitr::opts_knit$set(root.dir = "~/R_Projects/VCAS1_prelim_efficiency")
```

# Import the raw data - Point to the correct raw data directory

```{r}
raw_data_dir <- "data/processed/imported_efficiency_stats.RData"
```

## Import Subject Data

```{r}
load(raw_data_dir)
```

# Pick your patients

```{r}
# These patients are excluded from the procedural analysis set 
vt_efficiency_analyses_exclusions <- c(
  "1-001", # Procedural failure, in safety analysis but excluded from efficiency analysis set
  "1-011", # Procedural failure, in safety analysis but excluded from efficiency analysis set
  "1-021", # Screen failure, had RV/LV voltage mapping and no ablation
  "3-007" # Screen failure, discovered to be ICD dependent on the table 
)
```


```{r}
cleaned_dat_list <- lapply(dat_list, function(x)
  x |> filter(!(subjectId %in% vt_efficiency_analyses_exclusions)))

cleaned_res <- cleaned_res |>
  filter(!(subjectId %in% vt_efficiency_analyses_exclusions))
```

See bottom of the sheet for where we "pick" our patients even more! There has to be some filtering done AFTER we merge in the VCAS-I Index procedure data

# Label data according to studyid

```{r}
cleaned_res <- cleaned_res |>
  mutate(studyID = if_else(
    etiology == "idiopathic", "VCAS-II", "VCAS-I"
  ))
```

# Change Column Types

## Clean Column Names

```{r}
cleaned_headers_list <- headers_list |> clean_names()
cleaned_dat_list <- cleaned_dat_list |> clean_names()
```

# Grab numberapplications from GLG data

```{r}
cleaned_res <- full_join(
  cleaned_dat_list$dat_f14_index_procedure_for_vt |>
    select(subjectId, numberapplications, numberlesions),
  cleaned_res
)
```

# Calculate derived statistics

### Calculate timestamps

## Calculate time data

mapping duration
Total procedure duration = "vascular access" to "removal of study catheter from body"
Procedure time = "catheter in for mapping" to "removal of study catheter from body"
FieldForce catheter time in body
Ablation time - retrieve from our separate code

```{r}
cleaned_dat_list$dat_f14_index_procedure_for_vt <- cleaned_dat_list$dat_f14_index_procedure_for_vt |>
  mutate(
    # Convert data from character to time
    ablationcatheteroutofbodytime = hm(ablationcatheteroutofbodytime),
    ablationcatheterenterbodytime = hm(ablationcatheterenterbodytime),
    # Compute time differences
    ablationcathetertime = as.numeric(as.duration(ablationcatheteroutofbodytime - ablationcatheterenterbodytime),"minutes")
  ) 
```

### Use GLG timestamps for sites where we don't have footage to calculate dwell time

```{r}
# Identify all the non-Homolce VCAS data AND/OR Homolce data that we want to use GLG stuff for instead
id_to_replace <- cleaned_res |>
  # We want to exclude the homolka data except for 1-010, we do want to use GLG to impute for 1-010
  filter(!str_starts(subjectId, "1-") | subjectId == "1-010") |> select(subjectId)

# These are the non-Homolce data for dwell time from GLG
dwell_to_insert <- 
  cleaned_dat_list$dat_f14_index_procedure_for_vt |>
  filter(subjectId %in% id_to_replace$subjectId) |>
  # We derive ablationcathetertime in the procedural summaries workflow
  rename(dwell_time_min = ablationcathetertime) |>
  mutate(dwell_time = dwell_time_min * 60) |>
  select(subjectId, dwell_time_min, dwell_time)

# Iteratively replace the cleaned_res dwell_time data with the one from GLG from the non-Homolce
for(i in seq_along(1:nrow(id_to_replace))) {
  tmp <- id_to_replace[i,"subjectId"]
  cleaned_res[cleaned_res$subjectId == paste(tmp[[1]]),"dwell_time"] <- dwell_to_insert[dwell_to_insert$subjectId == paste(tmp[[1]]),"dwell_time"]
}
```

## Ablation Time for 1-009

We don't have footage or ablation timestamps for 1-009 so we're going to calculate ablation time manually

```{r}
# We know that the first application was delivered at "11:49:56" and that's corroborated bewteen footage and Michelle's hand taken records. So we will calculate difference between first application and catheter out time
#11:49:56
tmp <- cleaned_res |>
  filter(subjectId == "1-009") |>
  mutate(rove_time = as.numeric(cleaned_dat_list$dat_f14_index_procedure_for_vt[cleaned_dat_list$dat_f14_index_procedure_for_vt$subjectId == "1-009", "ablationcatheteroutofbodytime"][[1]] - lubridate::as.period(hms::as_hms("11:49:56"))))

cleaned_res <- cleaned_res |>
  filter(subjectId != "1-009") |>
  rbind(tmp)
```

### Do the math

```{r}
#app_per_rove is how many applications we can get in within a minute of rove_time
cleaned_res <- cleaned_res |>
  mutate(
    dwell_time_min = dwell_time / 60,
    rove_time_min = rove_time / 60,
    app_per_rove = numberapplications / rove_time_min, 
    irrigation_delivered = (dwell_time / 60) * 6,
    energy_on_time = case_when(
      therapy == "15kv_20x5_uni"
      ~ (numberapplications * 150) / 1000,
      therapy == "15kv_3x12_uni"
      ~ (numberapplications * 150) / 1000
    ),
    across(where(is.numeric), \(x) round(x, 2))
  ) 
```

## Convert to factor

```{r}
cleaned_res <- cleaned_res |>
  mutate(subjectId = factor(subjectId))
```

```{r}
cleaned_dat_list <-
  lapply(cleaned_dat_list,
         function(x)
           x |> mutate(subjectId = factor(subjectId, levels = levels(cleaned_res$subjectId)))
  )
```

# Pick your patients

This is currently at the top but it's likely redundant and we really need it here 

```{r}
#cleaned_dat_list <- lapply(dat_list, function(x)
#  x |> filter(
#         subjectId != "1-021", #screen failure
#         subjectId != "1-011"
#         ))

#cleaned_res <- cleaned_res |>
#  filter(
#         subjectId != "1-021", #screen failure
#         subjectId != "1-011"
#         )
```

# Save cleaned data

```{r}
save(cleaned_res, 
     file = paste0("data/processed/working_efficiency_stats.RData"))
```

# Version and Package Details

```{r echo=F}
paste(version$version.string, version$nickname)
paste0(
  "RStudio Version ",
  rstudioapi::versionInfo()$version,
  " ",
  rstudioapi::versionInfo()$release_name
) # RStudio Version
subset(data.frame(sessioninfo::package_info()),
       attached == TRUE,
       c(package, loadedversion)) # Loaded Packages
```

# When were these files last rewritten?

```{r}
date()
```
