==> code/step_3_descriptive_stats.rmd <==
---
title: "VCAS-I Efficiency: Procedural Analysis Set"
author: "Jonathan Salas"
date: "Last Updated `r format(Sys.time(), '%d %B, %Y')`"
output: 
  html_document: 
    df_print: kable
    fig_height: 6
    keep_md: true
---

# Start with a clean slate

## Remove all objects

```{r echo = F, include=FALSE}
rm(list = ls(all.names = TRUE)) # Remove all starting code
```

## Detach Packages

```{r include=FALSE}
try(invisible(lapply(paste0('package:', names(sessionInfo()$otherPkgs)), detach, character.only=TRUE, unload=TRUE)), silent = T) # Detach Packages
```

```{r setup, include = FALSE}
library(gtsummary)
library(MASS)
library(ggplot2)
library(cowplot)
library(dplyr)
knitr::opts_chunk$set(echo = F)
knitr::opts_knit$set(root.dir = "~/R_Projects/VCAS1_prelim_efficiency")
```

# Import - Point to the correct raw data directory

```{r}
raw_data_dir <- "data/processed/working_efficiency_stats.RData"
```

## load Subject Data

```{r}
load(raw_data_dir)
```

# Descriptive Statistics

Analysis Plan:
- histoplots of continuous variables - as of 7/30/24 this was useless to do, shows nothing at all due to small sample
- loess plots of continuous variables; x-axis is each procedure aligned in termporal order
- box plots
- we do procedure time vs rove time "proc_vs_rove" 

## The numbers

```{r}
cleaned_res |>
  select(-c(subjectId, dwell_time, rove_time, 
            # Correct etiologies found in medical history, not the Rove time sheet
            etiology
            )) |>
  tbl_summary(
    by = studyID,
    type = where(is.numeric) ~ "continuous2",
    statistic = list(
      all_continuous2() ~ c("{mean} ({sd})", "{median} [{p25}-{p75}]", "{min}, {max}"),
      all_categorical() ~ "{n} ({p}%)",
      missing_text = "NA"
    )
  ) |>
  add_n() |>
  add_overall() |>
  as_gt()
```

### VCAS-I Only

```{r}
cleaned_res |>
  select(-c(subjectId, dwell_time, rove_time, 
            # Correct etiologies found in medical history, not the Rove time sheet
            etiology
            )) |>
  filter(studyID == "VCAS-I") |>
  tbl_summary(
    by = studyID,
    type = where(is.numeric) ~ "continuous2",
    statistic = list(
      all_continuous2() ~ c("{mean} ({sd})", "{median} [{p25}-{p75}]", "{min}, {max}"),
      all_categorical() ~ "{n} ({p}%)",
      missing_text = "NA"
    )
  ) |>
  add_n() |>
  as_gt()
```

# VCAS-I Plots

## Make Mapping Data

Make temporary mapping data.

```{r}
# Make a procedure number column which is the patient subjectId's numbered sequentially
vt_temp_map_dat <- cleaned_res |>
  filter(studyID == "VCAS-I") |>
  mutate(proc_num = seq_along(subjectId)) 
```

## Irrigation Delivered

#### Line Plot

```{r, warning = FALSE,message = FALSE,error = FALSE}
plot <- ggplot(data = vt_temp_map_dat, aes(x = proc_num, y = irrigation_delivered)) +
  geom_point(size = 10) +
  geom_smooth(method = "rlm", se = F) +
  scale_x_continuous(labels = vt_temp_map_dat$proc_num, breaks = vt_temp_map_dat$proc_num) +
  labs(x = "Procedure Number", y = "Irrigation Delivered (mL)", caption = "Catheter Dwell Time * 6 mL/min") +
  coord_cartesian(ylim = c(0, 1500)) +
  theme_half_open()

plot

ggsave("irrigation_delivered_trend.png",
       plot = plot + theme_half_open(font_size = 34), 
       path = "figures",
       scale = 1,
       width = 28,
       height = 21,
       units = "cm",
       dpi = "retina",
       limitsize = F)
```

#### Box Plots

```{r, warning=FALSE,message=FALSE,error=FALSE}
plot <- ggplot(data = vt_temp_map_dat, aes(x = "", y = irrigation_delivered)) +
  geom_boxplot(outliers = F) +
  geom_point(
    size = 10,
    #    aes(color = subjectId),
    position = position_jitter(seed = 42494, width = .1),
    shape = 21,
    fill = "white"
  ) +
  xlab("") +
  ylab("Irrigation Delivered (mL)") +
  theme_half_open()  +
  scale_color_discrete(guide = "none") +
  scale_shape_discrete(guide = "none") +
  coord_cartesian(ylim = c(0, 1500))
#  theme(axis.text = element_text(size = 40, face = "bold"),
#        axis.title = element_text(size = 24,face ="bold"))
#  scale_y_continuous(position = "right")

plot

ggsave(
  filename = "irrigation_delivered_boxplot.png",
  plot = plot + theme_half_open(font_size = 34),
  path = "figures",
  scale = 1,
  width = 12,
  height = 21,
  units = "cm",
  dpi = 300,
  limitsize = F
)
```

## Ablation Time

#### Line Plot

```{r, warning = FALSE,message = FALSE,error = FALSE}
plot <- ggplot(data = vt_temp_map_dat, aes(x = proc_num, y = rove_time_min)) +
  geom_point(size = 10, 
             #aes(color = factor(proc_num))
             ) +
  geom_smooth(method = "rlm", se = F) +
  scale_x_continuous(labels = vt_temp_map_dat$proc_num, breaks = vt_temp_map_dat$proc_num) +
  xlab("Procedure Number") +
  ylab("Ablation Time (min)") +
  coord_cartesian(ylim=c(0,80)) +
  theme_half_open() 

plot

ggsave("ablation_time_trend.png",
       plot = plot + theme_half_open(font_size = 34),
       path = "figures",
       scale = 1,
       width = 28,
       height = 21,
       units = "cm",
       dpi = "retina",
       limitsize = F)
```

#### Box Plots

```{r, warning = FALSE,message = FALSE,error = FALSE}
plot <- ggplot(data = vt_temp_map_dat, aes(x = "", y = rove_time_min)) +
  geom_boxplot(outliers = F) +
  geom_point(
    size = 10,
#    aes(color = factor(proc_num)),
    position = position_jitter(seed = 42494, width = .1),
    shape = 21,
    fill = "white"
  ) +
  labs(x = "", y = "Transpired Ablation Time (min)") +
  theme_half_open() +
  coord_cartesian(ylim = c(0, 80)) +
  theme(legend.position = "none")

plot

ggsave("ablation_time_boxplot.png",
       plot = plot + theme_half_open(font_size = 34),
       path = "figures",
       scale = 1,
       width = 12,
       height = 21,
       units = "cm",
       dpi = "retina",
       limitsize = F)
```

### Energy On Time

#### Box Plots

```{r, warning = FALSE,message = FALSE,error = FALSE}
plot <- ggplot(data = vt_temp_map_dat, aes(x = "", y = energy_on_time)) +
  geom_boxplot(outliers = F) +
  geom_point(
    size = 10,
#    aes(color = factor(proc_num)),
    position = position_jitter(seed = 42494, width = .1),
    shape = 21,
    fill = "white"
  ) +
  labs(x = "", y = "Energy-On Time (s)") +
  theme_half_open() +
  coord_cartesian(ylim = c(0, 30)) +
  theme(legend.position = "none")
#  scale_y_continuous(position = "right")

plot

ggsave("energy_on_time_boxplot.png",
       plot = plot + theme_half_open(font_size = 34),
       path = "figures",
       scale = 1,
       width = 18,
       height = 21,
       units = "cm",
       dpi = "retina",
       limitsize = F)
```

# VCAS-II Plots

## Make Mapping Data

Make temporary mapping data.

```{r}
# Make a procedure number column which is the patient subjectId's numbered sequentially
pvc_temp_map_dat <- cleaned_res |>
  filter(studyID == "VCAS-II") |>
  mutate(proc_num = seq_along(subjectId)) 
```

## Ablation Time

#### Line Plot

```{r}
plot <- ggplot(data = pvc_temp_map_dat, aes(x = proc_num, y = rove_time_min)) +
  geom_point(size = 10, 
#             aes(color = factor(proc_num))
             ) +
  geom_smooth(span = 2, se = F) +
  scale_x_continuous(labels = pvc_temp_map_dat$proc_num, breaks = pvc_temp_map_dat$proc_num) +
  labs(x = "Procedure Number", y = "Transpired Ablation Time (min)", title = "VCAS-II") +
  coord_cartesian(ylim=c(0,10)) +
  theme_half_open() 
#  theme(axis.text = element_text(size=40, face = "bold"),
#        axis.title = element_text(size=24,face = "bold")) 

plot

ggsave("pvc_ablation_time_trend.png",
       plot = plot + theme_half_open(font_size = 34),
       path = "figures",
       scale = 1,
       width = 28,
       height = 21,
       units = "cm",
       dpi = "retina",
       limitsize = F)
```

#### Box Plots

```{r, warning = FALSE,message = FALSE,error = FALSE}
plot <- ggplot(data = pvc_temp_map_dat, aes(x = "", y = rove_time_min)) +
  geom_boxplot(outliers = F) +
  geom_point(
    size = 10,
#    aes(color = factor(proc_num)),
    position = position_jitter(seed = 42494, width = .1),
    shape = 21,
    fill = "white"
  ) +
  labs(x = "", y = "Transpired Ablation Time (min)") +
  theme_half_open() +
  coord_cartesian(ylim = c(0,10)) +
  theme(legend.position = "none")
#  scale_y_continuous(position = "right")

plot 

ggsave("pvc_ablation_time_boxplot.png",
       plot = plot + theme_half_open(font_size = 34),
       path = "figures",
       scale = 1,
       width = 18,
       height = 21,
       units = "cm",
       dpi = "retina",
       limitsize = F)
```

# Version and Package Details

```{r echo = F}
paste(version$version.string, version$nickname)
paste0(
  "RStudio Version ",
  rstudioapi::versionInfo()$version,
  " ",
  rstudioapi::versionInfo()$release_name
) # RStudio Version
subset(data.frame(sessioninfo::package_info()),
       attached == TRUE,
       c(package, loadedversion)) # Loaded Packages
```

# When were these files last rewritten?

```{r}
date()
```
